+++ src/router/index.js
@@ -1,0 +1,6 @@
+const productRouter =require('./productRouter');
+const userRouter = require('./userRouter');
+module.exports ={
+    productRouter,
+    userRouter
+}
