+++ src/router/userRouter.js
@@ -1,0 +1,6 @@
+const express =require('express');
+const userRouter = express();
+userRouter.get('/',(req,res)=>{
+res.json([])
+})
+module.exports = userRouter
